pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dockcadd"
version = "3.0.0"
description = "DockCADD-v3: automated end-to-end molecular docking workflow built around RDKit, P2Rank, Open Babel, and AutoDock Vina."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Karim El Mehdi" }
]
dependencies = [
  "biopython",
  "numpy",
  "openpyxl",
  "pandas",
  "pubchempy",
  "rdkit-pypi",
  "scipy",
  "tqdm",
]

[project.urls]
Homepage = "https://github.com/mehdikariim/DockCADD-v2"
DOI = "https://doi.org/10.1016/j.sciaf.2025.e02581"

[project.scripts]
dockcadd = "dockcadd.cli:main"

[tool.setuptools]
packages = ["dockcadd", "src", "caddock"]
